🚚 Rename: 폴더 구조 변경 common 하위 폴더 생성
