+++ src/App.tsx
@@ -1,6 +1,6 @@
 import { RouterProvider } from 'react-router-dom'
 import router from './routers/router'
-import Wrap from './components/common/Wrap'
+import Wrap from './components/common/wrap/Wrap'
 
 const App = (): React.ReactElement => {
   return (
