+++ src/pages/Index.tsx
@@ -1,6 +1,6 @@
 import image from '/images/mainImg.png'
 import styles from '@/styles/pages/index.module.scss'
-import TopLogo from '@/components/common/TopLogo'
+import TopLogo from '@/components/common/topLogo/TopLogo'
 import MainRouletteButton from '@/components/rouletteButton/MainRouletteButton'
 import GroupRouletteList from '@/components/rouletteButton/GroupRouletteList'
